refactor(server): split webhook alert delivery
